Remove size_t support from IPDL messages (1525199).

+++ dom/plugins/ipc/PluginModuleParent.cpp
@@ -2411,7 +2411,7 @@ PluginModuleParent::ProcessRemoteNativeEventsInInterruptCall()
 bool
 PluginModuleParent::RecvPluginShowWindow(const uint32_t& aWindowId, const bool& aModal,
                                          const int32_t& aX, const int32_t& aY,
-                                         const size_t& aWidth, const size_t& aHeight)
+                                         const double& aWidth, const double& aHeight)
 {
     PLUGIN_LOG_DEBUG(("%s", FULLFUNCTION));
 #if defined(XP_MACOSX)
